// jslib/models/FantasyLeagueSnapshot.ts
/**
 * Prop Odds API
 *  The Prop Odds API allows you to query player and game prop odds from different bookies. 
 *
 * OpenAPI spec version: 0.9
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { FantasyBookMultiMarket } from '../models/FantasyBookMultiMarket';
import { HttpFile } from '../http/http';

export class FantasyLeagueSnapshot {
    'league': string;
    'fantasyBooks': Array<FantasyBookMultiMarket>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "league",
            "baseName": "league",
            "type": "string",
            "format": ""
        },
        {
            "name": "fantasyBooks",
            "baseName": "fantasy_books",
            "type": "Array<FantasyBookMultiMarket>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return FantasyLeagueSnapshot.attributeTypeMap;
    }

    public constructor() {
    }
}