// jslib/models/FantasySnapshot.ts
/**
 * Prop Odds API
 *  The Prop Odds API allows you to query player and game prop odds from different bookies. 
 *
 * OpenAPI spec version: 0.9
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { FantasyBook } from '../models/FantasyBook';
import { HttpFile } from '../http/http';

export class FantasySnapshot {
    'league': string;
    'fantasyBooks': Array<FantasyBook>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "league",
            "baseName": "league",
            "type": "string",
            "format": ""
        },
        {
            "name": "fantasyBooks",
            "baseName": "fantasy_books",
            "type": "Array<FantasyBook>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return FantasySnapshot.attributeTypeMap;
    }

    public constructor() {
    }
}