// jslib/models/Game.ts
/**
 * Prop Odds API
 *  The Prop Odds API allows you to query player and game prop odds from different bookies. 
 *
 * OpenAPI spec version: 0.9
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Game {
    'gameId': string;
    'awayTeam'?: string;
    'homeTeam'?: string;
    'startTimestamp': string;
    'participants': Array<number>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "gameId",
            "baseName": "game_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "awayTeam",
            "baseName": "away_team",
            "type": "string",
            "format": ""
        },
        {
            "name": "homeTeam",
            "baseName": "home_team",
            "type": "string",
            "format": ""
        },
        {
            "name": "startTimestamp",
            "baseName": "start_timestamp",
            "type": "string",
            "format": ""
        },
        {
            "name": "participants",
            "baseName": "participants",
            "type": "Array<number>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Game.attributeTypeMap;
    }

    public constructor() {
    }
}