// jslib/models/GamesInfo.ts
/**
 * Prop Odds API
 *  The Prop Odds API allows you to query player and game prop odds from different bookies. 
 *
 * OpenAPI spec version: 0.9
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Game } from '../models/Game';
import { HttpFile } from '../http/http';

export class GamesInfo {
    'league': string;
    'date': string;
    'games': Array<Game>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "league",
            "baseName": "league",
            "type": "string",
            "format": ""
        },
        {
            "name": "date",
            "baseName": "date",
            "type": "string",
            "format": ""
        },
        {
            "name": "games",
            "baseName": "games",
            "type": "Array<Game>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GamesInfo.attributeTypeMap;
    }

    public constructor() {
    }
}