// jslib/models/Markets.ts
/**
 * Prop Odds API
 *  The Prop Odds API allows you to query player and game prop odds from different bookies. 
 *
 * OpenAPI spec version: 0.9
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Market } from '../models/Market';
import { HttpFile } from '../http/http';

export class Markets {
    'gameId': string;
    'markets': Array<Market>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "gameId",
            "baseName": "game_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "markets",
            "baseName": "markets",
            "type": "Array<Market>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Markets.attributeTypeMap;
    }

    public constructor() {
    }
}