// jslib/models/SportsBooks.ts
/**
 * Prop Odds API
 *  The Prop Odds API allows you to query player and game prop odds from different bookies. 
 *
 * OpenAPI spec version: 0.9
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { SportsBook } from '../models/SportsBook';
import { HttpFile } from '../http/http';

export class SportsBooks {
    'gameId': string;
    'sportsbooks': Array<SportsBook>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "gameId",
            "baseName": "game_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "sportsbooks",
            "baseName": "sportsbooks",
            "type": "Array<SportsBook>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return SportsBooks.attributeTypeMap;
    }

    public constructor() {
    }
}