// jslib/models/Usage.ts
/**
 * Prop Odds API
 *  The Prop Odds API allows you to query player and game prop odds from different bookies. 
 *
 * OpenAPI spec version: 0.9
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Usage {
    'plan': string;
    'used': number;
    'limit': number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "plan",
            "baseName": "plan",
            "type": "string",
            "format": ""
        },
        {
            "name": "used",
            "baseName": "used",
            "type": "number",
            "format": ""
        },
        {
            "name": "limit",
            "baseName": "limit",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Usage.attributeTypeMap;
    }

    public constructor() {
    }
}